Not allowed to use result outside of transaction
The results used within transactional functions were previously
automatically fetched when transaction was closed which made it
possible to use results outside of the transaction boundary, this is by
design not allowed. Could cause a lot of fetching and allocations that
might or might not be needed.

Utility functions to collect and get a single record has been changed so
that they will not build if used directly on transaction function
returns. New utility functions has been added so that same number of
lines is needed but not breaking.