tests: Moved stress tests to a separate console app
Rewrote stress tests from being based on test framework to being a
console app. But more important, it doesn't start it's own database
instance but rely on a database being started outside of the
application. This makes it possible to use for Aura drop tests as well
as running within testkit.