Fix ownership issue of byte slice used as a buffer
This change should fix #183. It appears that when we read a bolt message
into the buffer, accounting for dechunking, a reference to that byte
slice is then passed to the hydrator. However, since it's a reference to
the same underlying buffer using by the "incoming" struct for processing
bolt messages on the stream, it's possible data corruption can occur.