Remove discarded stream
The stream will not be readable after it has been discarded.

This could cause applications to get an error that they didn't
previously got IF they use results outside of transaction boundary
(which is not allowed) BUT even if they previously didn't see an error
they actually didn't get the correct result.