Introduce `context.Context`-based APIs
The driver now exposes a set of APIs, accessible via `neo4j
.NewDriverWithContext`.

Most of these APIs now accept an instance of `context.Context` as a first
parameter. This allows users to control how long the execution of each API can
take at most, through the built-in `context.WithTimeout` or
`context.WithDeadline` functions. They can also manually cancel an operation,
thus aborting any underlying network operation.

The core of these network operations are controlled with two new utilities:
 - `racing_io.RacingReader`
 - `racing_io.RacingWriter`
These new types allow to race any read or write operation against a given
context. They are used whenever a chunk of data is sent to or received from a
Bolt server.

Moreover, the initial TCP connection establishment and TLS handshake are now
context-aware.

The intersection of:
 - the existing timeouts `neo4j.ConnectionAcquisitionTimeout`
and `neo4j.SocketConnectTimeout`
 - connection read timeouts hints (enabled via the 4.3+ server setting `dbms
 .connector.bolt.connection_keep_alive_for_requests`)
 - user-provided `context.Context`s
is far from trivial and has been thoroughly documented in the `neo4j.Config`
struct. Users are recommended to not mix custom timeouts and user-provided
contexts.

The usage of connection acquisition timeout is now in accordance with the way it
 is documented: the timeout now includes the whole execution time of a
 connection borrowing from the pool. Prior to this commit, there could be
 situations where a connection would successfully but slowly be established and
 returned as a borrowed connection, even though it had taken more than the
 connection acquisition timeout allowed.

The existing APIs, accessible via `neo4j.NewDriver` are now considered
deprecated. They will be removed in 6.0. They delegate to the context-aware
connection layer of the driver (with non-cancelling, non-terminating contexts).