Do not reset ready connections
Ready connections are in this state because one of the following
events occurred:

 1. they successfully authenticated after initial handshake
 2. their transaction completed
 3. there are no more active streams

In the first case, there is nothing to reset.

In the second case, the connection returns to the pool right after.

The third case is less clear. After some manual tests, the lack
of reset should be OK, since an active connection cannot be used by
several sessions at once and a single session cannot run multiple
explicit transaction at once (and are configured with a single
database name in their lifetime).