Close underlying connection when context has terminated
This allows the Bolt server to ditch the transaction closely after
the context terminates.

Before this, the only way to have the Bolt server release
transactions was to close the entire connection pool (which is
done when closing the driver). That's because the connection is
marked as dead, which means the pool will not act on it upon
return.