Warn of the impact of early context terminations
Since 78c1c753a111bc0d1d3a4d26d36411049d863258, an early context
termination results in the active connection's socket to be closed.

This potentially valid connection will be marked as dead and will
not be a candidate for future reuse.

Moreover, the cleanup of idle connections that are too old may not
even occur if the context has terminated. This results in the pool
gradually filling up with potentially bad connections.