Introduce message queue to better correlate requests and responses
This is currently only applied to Bolt 5.x, but can be ported to
older protocol connections as well.

This removes some fragile heuristics on response correlation
(see e.g. success#isResetResponse).

This also paves the way for a cleaner implementation of re-auth.

Big kudos to @robsdedude for guiding through the design principles
of this solution.